+++ backend/src/users/users.module.ts
@@ -3,10 +3,12 @@ import { MongooseModule } from '@nestjs/mongoose';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { User, UserSchema } from './users.schema';
+import { MapConfigModule } from '../map-config/map-config.module';
 
 @Module({
   imports: [
     MongooseModule.forFeature([{ name: User.name, schema: UserSchema }]),
+    MapConfigModule,
   ],
   controllers: [UsersController],
   providers: [UsersService],